\name{data}
\docType{data}
\alias{data}

\title{Trait data}
\usage{
  data
}
\description{
  Simulated trait records data for the package P-SIMEX. 
  It contains information on each individual's trait and covariates.
             
}
\format{
"sex" "f_inb" "animal.id" "id" "year" "animal" "y"  
  \tabular{ll}{
    sex:\tab Sex of the individual\cr
    f_inb:\tab Inbreeding coefficient of the individual 
    (it can also be calculated from the pedigree)\cr
    id:\tab Individual's id \cr
    animal.id:\tab Individual's id (duplicate for animal model) \cr
    animal:\tab Individual's id (duplicate for animal model) \cr
    year:\tab Generation number of the individual in the pedigree
    (it can also be extracted from the pedigree)\cr
    y:\tab Individual's trait record\cr
  
 
  
}
}

\keyword{datasets}
\keyword{PSIMEX}