\name{pedigree}
\docType{data}
\alias{pedigree}

\title{Pedigree data}
\usage{
pedigree
}
\description{
Simulated pedigree structure for the package P-SIMEX. 
It contains information on each individual's parents, 
as well its sex and generation.
             
}
\format{
  
  \tabular{ll}{
    id:\tab Individual's id \cr
    father:\tab Individual's father's id\cr
    mother:\tab Individual's mother's id\cr
    sex:\tab Sex of the individual\cr
    generation:\tab Generation number of the individual in the pedigree \cr
    
  }
}

\keyword{datasets}
\keyword{PSIMEX}