\name{PSIMEX-package}
\alias{PSIMEX-package}
\alias{PSIMEX}
\docType{package}
\title{
\packageTitle{PSIMEX}
}
\description{
\packageDescription{PSIMEX}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PSIMEX}
\packageIndices{PSIMEX}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PSIMEX}

Maintainer: \packageMaintainer{PSIMEX}
}
